Floats

+++ include/motion_control/motor.h
@@ -1,24 +1,22 @@
 #ifndef MOTION_MOTOR_H
 #define MOTION_MOTOR_H
 
-#include <cinttypes>
-
 namespace MotionControl {
 
 // Motor Interface
 //
 // A Motor's output voltage is represented by an integer ranging from
-// the minimum possible voltage INT16_MIN, to zero voltage 0, to the
-// maximum possible voltage INT16_MAX.
+// the minimum possible voltage -1.0, to zero voltage 0.0, to the
+// maximum possible voltage +1.0.
 //
 class Motor
 {
 public:
     // Returns the most recently set motor voltage.
-    virtual int_fast16_t voltage() const = 0;
+    virtual float voltage() const = 0;
 
     // Sets the motor voltage. Returns success.
-    virtual bool voltage(int_fast16_t value) = 0;
+    virtual bool voltage(float value) = 0;
 };
 
 }
